requirements.txt
gradio
pdf2docx
PyMuPDF
python-docx

